Revamped Ball Subsystem; Converted incorrect relative changes-based servo coding to absolute changes-based servo coding; Included Ball into Robot.cpp/Robot.h; Included cases for release All and Toggle; Changed cases for Release One to give time for servos to close.

+++ src/main/include/RobotControlData.h
@@ -31,7 +31,7 @@ struct ElevatorOutput{
 };
 
 struct ClawInput{
-    int ClawState;
+    double ClawState;
     int WiggleState;
 };
